pip_requirements.txt
aniso8601==8.0.0
astroid==2.3.3
atomicwrites==1.4.0
attrs==19.3.0
autopep8==1.4.4
backcall==0.1.0
bleach==3.1.4
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cloudpickle==1.2.2
colorama==0.4.3
cycler==0.10.0
cytoolz==0.10.1
dask==2.9.1
decorator==4.4.1
defusedxml==0.6.0
entrypoints==0.3
Flask==1.1.2
Flask-Cors==3.0.8
Flask-RESTful==0.3.8
future==0.18.2
idna==2.9
imageio==2.6.1
importlib-metadata==1.5.0
imutils==0.5.3
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
mistune==0.8.4
mkl-service==2.3.0
more-itertools==8.2.0
nbconvert==5.6.1
nbformat==5.0.6
networkx==2.4
notebook==6.0.3
numpy==1.17.3
olefile==0.46
packaging==20.3
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
Pillow==7.0.0
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==3.0.4
py==1.8.1
pycodestyle==2.5.0
pyglet==1.4.9
Pygments==2.6.1
pylint==2.4.4
pyparsing==2.4.6
pyrsistent==0.16.0
pytest==5.4.1
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
pywin32==227
pywinpty==0.5.7
pyzmq==18.1.1
qtconsole==4.7.4
QtPy==1.9.0
requests==2.23.0
rope==0.16.0
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.3.2
Send2Trash==1.5.0
sip==4.19.13
six==1.13.0
terminado==0.8.3
testpath==0.4.4
toolz==0.10.0
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.12.1
zipp==3.1.0
